requirements.txt
-i https://pypi.org/simple
anyio==3.7.1; python_version >= '3.7'
attrs==23.1.0; python_version >= '3.7'
aws-cdk-lib==2.100.0
aws-cdk.asset-awscli-v1==2.2.200; python_version ~= '3.7'
aws-cdk.asset-kubectl-v20==2.1.2; python_version ~= '3.7'
aws-cdk.asset-node-proxy-agent-v6==2.0.1; python_version ~= '3.7'
black==23.9.1
cattrs==23.1.2; python_version >= '3.7'
certifi==2023.7.22; python_version >= '3.6'
click==8.1.7; python_version >= '3.7'
constructs==10.3.0
coverage[toml]==7.3.2; python_version >= '3.8'
h11==0.14.0; python_version >= '3.7'
httpcore==0.18.0; python_version >= '3.8'
httpx==0.25.0
idna==3.4; python_version >= '3.5'
importlib-resources==6.1.0; python_version >= '3.8'
iniconfig==2.0.0; python_version >= '3.7'
jsii==1.90.0; python_version ~= '3.7'
mypy-extensions==1.0.0; python_version >= '3.5'
packaging==23.2; python_version >= '3.7'
pathspec==0.11.2; python_version >= '3.7'
platformdirs==3.11.0; python_version >= '3.7'
pluggy==1.3.0; python_version >= '3.8'
publication==0.0.3
pytest==7.4.2
pytest-cov==4.1.0
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.0; python_version >= '3.7'
typeguard==2.13.3; python_full_version >= '3.5.3'
typing-extensions==4.8.0; python_version >= '3.8'
annotated-types==0.6.0; python_version >= '3.8'
fastapi==0.103.2
mangum==0.17.0
pydantic==2.4.2; python_version >= '3.7'
pydantic-core==2.10.1; python_version >= '3.7'
starlette==0.27.0; python_version >= '3.7'
uvicorn==0.23.2
